refactor code moving more content to the utils.py
